Complete astar algorithm with a simple heuristic.
Make the program significantly faster.
Delete block.py and Block class. Now the states are represented as a
dictionary with keys the block (e.g. A, B etc) and with values a
dictionary describing its position (clear, on, ontable).
Rename world.py to state.py and class World to class WorldState.